# specs/05-spring-openapi.yaml
openapi: 3.0.0
info:
  title: Spring Boot Main Backend API
  version: 1.0.0
paths:
...
  /api/v1/devices/{deviceId}/rfid-scans:
    post:
      summary: Receive UID and start verification
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                uid:
                  type: string
                scannedAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: OK
  /api/v1/devices/{deviceId}/heartbeat:
    post:
      summary: Device heartbeat
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/devices/{deviceId}/incidents:
    post:
      summary: Push button incident report
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                source:
                  type: string
                occurredAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: OK
...